pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "lobesync"
version = "0.1.0"
description = "Personal AI assistant for tasks, notes, memories, and checklists"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.89.0",
    "langgraph>=1.1.6",
    "sqlmodel>=0.0.38",
    "rich>=14.3.3",
    "python-dotenv>=1.2.2",
]

[project.scripts]
lobesync = "lobesync.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lobesync*"]
exclude = [".venv*", "__pycache__*"]
